requirements.txt
aiohttp==3.4.4
appnope==0.1.0
astroid==2.0.4
async-timeout==3.0.0
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==2.1.7
django-cors-headers==2.4.0
django-enumfields==0.10.1
django-filter==2.1.0
django-filters==0.2.1
django-jalali==3.1.0
django-rest-framework==0.1.0
django-rest-passwordreset==0.9.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
idna-ssl==1.1.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jdatetime==3.2.0
jedi==0.13.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
requests==2.20.1
six==1.11.0
sorl-thumbnail==12.5.0
sorl-thumbnail-serializer-field==0.2.1
telepot==12.7
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
whitenoise==4.1
wrapt==1.10.11
yarl==1.2.6
